api/requirements.txt
# ECL Facebook Fork
git+https://github.com/sitterfied/ECL-Facebook.git#egg=ecl-facebook

# Parse DateTime Master
git+https://github.com/bear/parsedatetime.git@863c066338db80edd4045626a196f4f9f9e7ae71

boto==2.19.0
celery==3.1.18
cssmin==0.2.0
delorean==0.4.5
django==1.8.2
django-annoying==0.8.0
django-autofixture==0.10.1
django-cors-headers==1.1.0
django-crispy-forms==1.4.0
django-extensions==1.2.5
django-favicon==0.1.1
django-filter==0.10.0
django-floppyforms==1.1
django-localflavor-us==1.1
django-model-utils==2.2
django-pipeline==1.3.16
django-pipeline-compass==0.1.4
django-redis-cache==0.10.0
django-redis-sessions==0.4.0
django-registration-redux==1.1
django-ses==0.7.0
django-smtp-ssl==1.0
django-storages==1.1.8
django-twilio==0.7.2
djangorestframework==3.2.3
djrill==1.3.0
drf-ujson==1.1
facebook-sdk==0.4.0
feedparser==5.1.3
gdata==2.0.18
geopy==1.8.1
guppy==0.1.10
hiredis==0.1.1
ipython==1.1.0
lxml==3.2.4
mock==1.0.1
nose==1.3.0
oauth2client==1.2
pillow==2.3.0
ply==3.4
psycopg2==2.4.5
pyatom==1.4
pyhamcrest==1.8.0
python-magic==0.4.6
pytz==2013.8
pyuploadcare==1.2.2
redis==2.8.0
requests==2.0.1
six==1.5.2
slimit==0.8.1
twilio==3.6.4
unipath==1.0
uwsgi==2.0.7
watchdog==0.6.0
